# DirtyNodesCache optimizations