--- a/src/Nethermind/Nethermind.Benchmark/Store/TrieStoreBenchmarks.cs
+++ b/src/Nethermind/Nethermind.Benchmark/Store/TrieStoreBenchmarks.cs
@@ -1,4 +1,4 @@
-﻿// SPDX-FileCopyrightText: 2023 Demerzel Solutions Limited
+// SPDX-FileCopyrightText: 2023 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
 using BenchmarkDotNet.Attributes;