This PR optimizes two call sites of `ConcurrentDictionary` in the `DirtyNodesCache` of `TrieStore`. These methods are called heavily when pruning and adding nodes.

## Benchmarks

BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22621.1265/22H2/2022Update/SunValley2)
12th Gen Intel Core i9-12900HK, 1 CPU, 20 logical and 14 physical cores
.NET SDK=7.0.100
  [Host]     : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2


|                                         Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|----------------------------------------------- |---------:|---------:|---------:|-------:|----------:|
| FindCachedOrUnknown_And_Delete (before) | 72.78 ns | 0.788 ns | 0.658 ns | 0.0076 |      96 B |
| FindCachedOrUnknown_And_Delete (after) | 63.21 ns | 1.045 ns | 0.926 ns | 0.0076 |      96 B |

## Changes

- remove `Remove` with native `TryRemove`
- `GetOrAdd` instead of two calls to the cache

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [x] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [ ] Yes
- [x] No

#### If yes, did you write tests?

- [ ] Yes
- [x] No