@Scooletz 2023-03-02
The memory calculations go wrong as atomicity is needed for good book keeping.